separate search bar and table into separate components * fixed hydration error due to no <tr> tag inside <thead> * added debounce to search input to prevent unnecessary filtering computation
